Restructured extension files
