+++ src/Template.Application/Extensions/DependencyInjection/DependencyInjection.cs
@@ -7,7 +7,7 @@ using Template.Application.IdentityServer.Providers;
 using Template.Application.Validation.Interfaces;
 using Template.Application.Validation.Services;
 
-namespace Template.Application;
+namespace Template.Application.Extensions.DependencyInjection;
 
 public static class DependencyInjection
 {
@@ -30,6 +30,7 @@ public static class DependencyInjection
         });
 
         services.AddValidatorsFromAssembly(assembly);
+
         services.AddMediatR(options =>
         {
             options.RegisterServicesFromAssembly(assembly);
