+++ src/Template.Infrastructure/Extensions/DependencyInjection/DependencyInjection.cs
@@ -4,11 +4,9 @@ using Template.Application.Email.Interfaces;
 using Template.Application.Identity.Interfaces;
 using Template.Domain.Common.Models;
 using Template.Infrastructure.Email.Services;
-using Template.Infrastructure.Identity.Extensions;
 using Template.Infrastructure.Identity.Services;
-using Template.Infrastructure.IdentityServer.Extensions;
 
-namespace Template.Infrastructure;
+namespace Template.Infrastructure.Extensions.DependencyInjection;
 
 public static class DependencyInjection
 {
